lint: fix linter problems, also renaming
