+++ src/r-bridge/lang-4.x/ast/index.ts
@@ -1,3 +1,3 @@
 export * from './model'
-export * from './parser/xml'
+export * from './parser/main'
 export { parseLog } from './parser/json/parser'
