+++ src/r-bridge/lang-4.x/ast/parser/main/index.ts
@@ -1,3 +1,2 @@
 export * from './data'
 export * from './hooks'
-export * from './input-format'
