+++ test/functionality/r-bridge/lang/ast/parse-snippets.ts
@@ -2,7 +2,7 @@ import { exprList, numVal } from '../../../_helper/ast-builder'
 import { assertAst, withShell } from '../../../_helper/shell'
 import { rangeFrom } from '../../../../../src/util/range'
 import { RType } from '../../../../../src/r-bridge'
-import { ensureExpressionList } from '../../../../../src/r-bridge/lang-4.x/ast/parser/xml/internal'
+import { ensureExpressionList } from '../../../../../src/r-bridge/lang-4.x/ast/parser/main/internal'
 
 describe('Parse larger snippets', withShell((shell) => {
 	describe('if-then, assignments, symbols, and comparisons', () => {
